Rename assignment to task for consistency and async calls
